requirements.txt
python-telegram-bot==20.7
sqlalchemy==2.0.23
alembic==1.13.1
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
loguru==0.7.2
pytest==7.4.3
pytest-asyncio==0.21.1
